quell type warning, cast enum to int
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@13741 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/DynamicColor.cxx
@@ -227,9 +227,9 @@ void DynamicColor::finalize()
   for (int c = 0; c < 4; c++) {
     sequenceParams& seq = channels[c].sequence;
     for (unsigned int i = 0; i < seq.count; i++) {
-      if ((int)seq.list[i] < colorMin) {
+      if ((int)seq.list[i] < (int)colorMin) {
 	seq.list[i] = colorMin;
-      } else if (seq.list[i] > colorMax) {
+      } else if ((int)seq.list[i] > (int)colorMax) {
 	seq.list[i] = colorMax;
       } else if ((c == 3) && (seq.list[i] == colorMid)) {
 	noAlphaSeqMid = false;
